sprint 3 fixed comments
