+++ src/components/App/App.module.css
@@ -1,0 +1,7 @@
+.headerText {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-top: 120px;
+  margin-bottom: 20px;
+}
