+++ src/pages/login.jsx
@@ -1,7 +1,6 @@
-import React, {useRef, useEffect} from "react";
+import React, {useEffect} from "react";
 import {useDispatch, Provider, useSelector } from 'react-redux';
-import AppHeader from "../components/AppHeader/AppHeader";
-import { Link, useHistory, useLocation, Redirect } from 'react-router-dom';
+import { Link, useHistory, useLocation } from 'react-router-dom';
 import { Input, PasswordInput, Button} from '@ya.praktikum/react-developer-burger-ui-components';
 import loginStyle from "./login.module.css";
 import {login} from '../services/actions/auth.js';
@@ -19,14 +18,18 @@ export function LoginPage() {
 
 function Login() {
     const history = useHistory();
-    const email = useRef(null);
     const dispatch = useDispatch();
     const [password, setPassword] = React.useState('');
+    const [email, setEmail] = React.useState('');
     const {isAuth} = useSelector(state => state.userInfo);
     const location = useLocation();
 
     const onChangePassword = (e) => {
         setPassword(e.target.value);
+    }
+
+    const onChangeEmail= (e) => {
+        setEmail(e.target.value);
     }
 
     useEffect(()=> {
@@ -40,25 +43,24 @@ function Login() {
         }
     }, [isAuth]);
 
-    const handleLogin = () => {
-        if (email.current.value.length !== 0 && password.length !== 0) {
-            dispatch(login(email.current.value, password));
-        } 
+    const handleLogin = (e) => {
+        e.preventDefault(true);
+        dispatch(login(email, password)); 
     }
 
 
     return (
-    <div>
+    <form onSubmit={handleLogin}>
         <div className={loginStyle.loginStyle}>
             <p className="text text_type_main-medium">Вход</p>
             <div className={loginStyle.marginTop}>
-                <Input type="email" placeholder="E-mail" ref={email}></Input>
+                <Input value={email} type="email" placeholder="E-mail" onChange={onChangeEmail}></Input>
             </div>
             <div className={loginStyle.marginTop}>
                 <PasswordInput value={password} onChange={onChangePassword}></PasswordInput>
             </div>
             <div className={loginStyle.marginTop}>
-                <Button type="primary" size="large" onClick={handleLogin}>
+                <Button type="primary" size="large">
                 Войти
                 </Button>
             </div>
@@ -71,6 +73,6 @@ function Login() {
                 </p>
             </div>
         </div>
-    </div>
+    </form>
     )
   }
