+++ src/pages/main.jsx
@@ -5,7 +5,7 @@ import BurgerIngredients from "../components/BurgerIngredients/BurgerIngredients
 import BurgerConstructor from "../components/BurgerConstructor/BurgerConstructor";
 import mainStyle from "./main.module.css";
 import { Provider } from 'react-redux';
-import {addConstructorIngredient, setCountIngredient, deleteConstructorIngredient, deleteCountIngredient} from '../services/actions/index.js';
+import {addConstructorIngredient, setCountIngredient, deleteConstructorIngredient, deleteCountIngredient} from '../services/actions/ingredients.js';
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
 import {store} from '../services/store.js';
